# openapi.yaml
openapi: 3.1.0
info:
  version: 1.0.0
  title: Http Server I/F
paths:
  /health:
    get:
      summary: Health Check
      operationId: health
    responses:
      200:
        description: no content
      default:
        description: unexpected error
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Error"
  /auth/signin:
    post:
      summary: signin
      description: signin an account
      operationId: postSignin
      requestBody:
        description: signin request
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SigninRequest"
      responses:
        "200":
          description: signin response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SigninResponse"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /account/{id}:
    get:
      summary: Returns account by id
      operationId: getAccount
      parameters:
        - name: id
          in: path
          description: ID of account to fetch
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: account response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      summary: Update account by id
      operationId: putAccount
      requestBody:
        description: Account to add to the store
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Account"
      parameters:
        - name: id
          in: path
          description: ID of account to fetch
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: account response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      summary: Delete account by id
      operationId: deleteAccount
      parameters:
        - name: id
          in: path
          description: ID of account to fetch
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: no content
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /account:
    post:
      summary: Creates a new account
      description: Creates a new account in the store. Duplicates are not allowed
      operationId: postAccount
      requestBody:
        description: Account to add to the store
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewAccount"
      responses:
        "200":
          description: account response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  schemas:
    Account:
      allOf:
        - $ref: "#/components/schemas/NewAccount"
        - required:
            - id
          properties:
            id:
              type: integer
              format: int64
              description: Unique id of the account
    NewAccount:
      required:
        - name
      properties:
        name:
          type: string
          description: Name of the account
    SigninRequest:
      required:
        - id
        - pass
      properties:
        id:
          type: integer
          format: int64
        pass:
          type: string
    SigninResponse:
      required: 
        - token
      properties:
        token:
          type: string
    Error:
      required:
        - message
      properties:
        message:
          type: string
          description: Error message